# Conflicts:
#	src/utils/checksumUtils.ts
#	src/utils/streamUtils.spec.ts
#	src/utils/streamUtils.ts

+++ src/utils/streamUtils.ts
@@ -1,11 +1,23 @@
 import { F_OK } from 'node:constants'
 import { createWriteStream, createReadStream, access } from 'node:fs'
-import { unlink } from 'node:fs/promises'
+import { stat, unlink } from 'node:fs/promises'
 import type { Readable } from 'node:stream'
 import { pipeline } from 'node:stream'
 
 export type ReadableProvider = {
-  createStream(): Readable
+  /**
+   * Guarantees to provide a new stream every time this is called, before `destroy` is invoked.
+   */
+  createStream(): Promise<Readable>
+
+  /**
+   * Returns size of the persisted content
+   */
+  getContentLength(): Promise<number>
+
+  /**
+   * Remove the persisted value. No new streams can be created after this is called
+   */
   destroy(): Promise<void>
 }
 
@@ -32,6 +44,14 @@ export class FsReadableProvider implements ReadableProvider {
     })
   }
 
+  async getContentLength(): Promise<number> {
+    if (!(await this.fileExists())) {
+      throw new Error(`File ${this.storageFile} was already deleted.`)
+    }
+    const stats = await stat(this.storageFile)
+    return stats.size
+  }
+
   async createStream(): Promise<Readable> {
     if (!(await this.fileExists())) {
       throw new Error(`File ${this.storageFile} was already deleted.`)
@@ -40,7 +60,12 @@ export class FsReadableProvider implements ReadableProvider {
     return createReadStream(this.storageFile)
   }
 
-  destroy(): Promise<void> {
+  async destroy(): Promise<void> {
+    if (!(await this.fileExists())) {
+      // nothing to do here
+      return
+    }
+
     return unlink(this.storageFile)
   }
 
@@ -67,4 +92,22 @@ export class FsReadableProvider implements ReadableProvider {
     await provider.persist(options.sourceReadable)
     return provider
   }
+}
+
+export function getReadableContentLength(readable: Readable): Promise<number> {
+  return new Promise((resolve, reject) => {
+    let size = 0
+    readable.on('data', (chunk) => {
+      if (typeof chunk === 'string' || Buffer.isBuffer(chunk)) {
+        size += Buffer.byteLength(chunk)
+      }
+    })
+    readable.on('end', () => {
+      resolve(size)
+    })
+    readable.on('error', (error) => {
+      /* c8 ignore next 1 */
+      reject(error)
+    })
+  })
 }
